Split the two files my changes pushed over the size guard
The CI size guard failed on presentation.test.ts (537 of 520) and
results.css (471 of 460). Both exceptions carried a note naming the split
to make next, so that is what this does rather than raising the caps.

The error-boundary styles become their own top-level stylesheet. They
replace the entire shell rather than styling anything inside it, so they
were never really part of results.css.

The value and CSV-serialisation tests become presentation.values.test.ts,
which is the "one file per helper family" the exception asked for.

Both caps are lowered to the new sizes so the reclaimed headroom does not
quietly refill.

Also drops `.at(-1)` from the mDNS test: it needs a newer lib target than
this project sets, and `tsc -b` had cached a pass over it.
