+++ apps/netscli-gui/src/App.css
@@ -2,6 +2,7 @@
 @import './styles/shell.css';
 @import './styles/forms.css';
 @import './styles/results.css';
+@import './styles/error-boundary.css';
 @import './styles/pcap.css';
 @import './styles/details.css';
 @import './styles/responsive.css';
