+++ apps/netscli-gui/src/styles/error-boundary.css
@@ -1,0 +1,41 @@
+/* The whole-window fallback when a render throws.
+
+   Its own file rather than part of results.css: this replaces the entire
+   shell rather than styling anything inside it, and results.css was at its
+   size cap with a note asking for exactly this kind of split. */
+
+/* Replaces the entire window when a render throws, so it carries its own
+   background rather than inheriting one from a shell that is no longer
+   mounted. */
+.error-boundary {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  align-items: flex-start;
+  min-height: 100vh;
+  padding: 32px;
+  color: var(--text-primary);
+  background: var(--bg-body);
+}
+
+.error-boundary h1 {
+  margin: 0;
+  font-size: 18px;
+}
+
+.error-boundary p {
+  max-width: 60ch;
+  margin: 0;
+  color: var(--text-secondary);
+}
+
+.error-boundary-detail {
+  max-width: 100%;
+  margin: 0;
+  padding: 10px 12px;
+  overflow-x: auto;
+  color: var(--red);
+  background: color-mix(in srgb, var(--red), transparent 92%);
+  border: 1px solid color-mix(in srgb, var(--red), transparent 60%);
+  border-radius: 4px;
+}
