+++ apps/netscli-gui/src/workspace/toolExecution.test.ts
@@ -108,17 +108,23 @@ describe('numeric form bounds', () => {
   it('clamps the mDNS timeout to the range its field declares', async () => {
     const netscli = await import('../services/netscli');
     const tab = createTab('mdns');
+    // Indexed rather than `.at(-1)`, which needs a newer lib target than
+    // this project sets.
+    const lastTimeout = () => {
+      const calls = vi.mocked(netscli.discoverMdns).mock.calls;
+      return calls[calls.length - 1]?.[0];
+    };
 
     tab.form.timeout_ms = '999999';
     await executeTool(tab, 'op-mdns-1', 64);
-    expect(vi.mocked(netscli.discoverMdns).mock.calls.at(-1)?.[0]).toBe(30000);
+    expect(lastTimeout()).toBe(30000);
 
     tab.form.timeout_ms = '1';
     await executeTool(tab, 'op-mdns-2', 64);
-    expect(vi.mocked(netscli.discoverMdns).mock.calls.at(-1)?.[0]).toBe(500);
+    expect(lastTimeout()).toBe(500);
 
     tab.form.timeout_ms = '3000';
     await executeTool(tab, 'op-mdns-3', 64);
-    expect(vi.mocked(netscli.discoverMdns).mock.calls.at(-1)?.[0]).toBe(3000);
+    expect(lastTimeout()).toBe(3000);
   });
 });
